Move release notes helpers to releaseNotes.ts
No other changes made.
